chore: changed folder structure
