+++ src/index.js
@@ -4,7 +4,7 @@ import App from "./App";
 import { BrowserRouter } from "react-router-dom";
 import { makeServer } from "./server";
 import "./index.css";
-import { store } from "./redux/store";
+import { store } from "./redux-management/store";
 import { Provider } from "react-redux";
 
 // Call make Server
