+++ src/redux-management/index.js
@@ -1,0 +1,44 @@
+export { logInUser, signUpUser } from "./slices/authSlice/authServices";
+export { setFoundUser } from "./slices/authSlice/authSlice";
+
+export {
+  addCommentToAPost,
+  deleteComment,
+  downVoteComment,
+  editComment,
+  getCommentsOfAPost,
+  upVoteComment,
+} from "./slices/postSlice/postServices/commentServices";
+export {
+  dislikePost,
+  likePost,
+} from "./slices/postSlice/postServices/likeServices";
+export {
+  createNewPost,
+  deletePost,
+  editPost,
+  getAllPosts,
+  getPostById,
+  getPostsByUsername,
+} from "./slices/postSlice/postServices/postServices";
+export {
+  setPostId,
+  setIsEdit,
+  setContent,
+  setCommentId,
+} from "./slices/postSlice/postSlice";
+
+export {
+  bookmarkPost,
+  getAllBookmarkedPosts,
+  removeBookmark,
+} from "./slices/userSlice/userServices/bookmarkServices";
+export {
+  followUser,
+  unFollowUser,
+} from "./slices/userSlice/userServices/followServices";
+export {
+  editUserProfile,
+  getAllUsers,
+  getUserDetails,
+} from "./slices/userSlice/userServices/userServices";
