+++ src/redux-management/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
-import postReducer from "./slices/postSlice";
-import authReducer from "./slices/authSlice";
-import userReducer from "./slices/userSlice";
+import postReducer from "./slices/postSlice/postSlice";
+import authReducer from "./slices/authSlice/authSlice";
+import userReducer from "./slices/userSlice/userSlice";
 
 export const store = configureStore({
   reducer: {
